show group colors using html in cell
This provides a better fix for bug #18 and is related to #9.
The basic approach taken is to update pdata with a div>span when a group is added. By
doing so, I can achieve the desired styling through css without external javascript (which was being reset by DT on sorting).


Former-commit-id: f982d772fcdd2b8da04ff1e3af373a280770090d